+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [v2.4.1](https://github.com/simvue-io/python-api/releases/tag/v2.4.1) - 2026-03-31
+
+- Moved to using `threading.Event` as termination trigger events and added deprecation notice for `multiprocessing.Event`.
+- Fixed creation of duplicate alerts.
+- Fixed case where alert specifications being `0` led to incorrect falsification.
+
 ## [v2.4.0](https://github.com/simvue-io/python-api/releases/tag/v2.4.0) - 2026-03-23
 
 - Added search for Simvue configuration within the parent file hierarchy of the current directory.